Allow reading extensions from yaml
The extensions consist of two addresses: (1) 'extension_file' includes the address of a script for any decorations of existing nodes/arcs/models objects; (2) 'new_classes' includes the new node name and the associated script address. It is noted that only the folder address is provided rather than the exact script address.